+++ packages/matrix-code/src/components/molecules/base-matrix-code/base-matrix-code.tsx
@@ -1,6 +1,6 @@
 import { BaseButton, BaseTypography } from '@mission-platform/components';
 import { IconCheck, IconCopy, IconDownload, IconImage } from '@mission-platform/icons';
-import { h, useEffect, useMemo, useRef, useState, type MpElement, type MpProperties } from '@mission-platform/forge';
+import { h, type MpElement, type MpProperties, useEffect, useMemo, useRef, useState } from '@mission-platform/forge';
 import { encodeMatrix, type MatrixSymbology } from '@mission-platform/matrix-code';
 
 import styles from './base-matrix-code.module.scss';